feat: add CorsTest component for CORS testing in development and update API base URL
